+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 
 * Broaden stray light search window to 31 in https://github.com/punch-mission/punchpipe/pull/217
+* Fixes to run the stray light model generation flow in https://github.com/punch-mission/punchpipe/pull/216
 
 ## Version 0.0.11: July 23, 2025